[Dashboard] Add dashboard model to improve module UI build process
