+++ dashboard/src/main/java/mx/yellowme/youst/dashboard/domain/Dashboard.kt
@@ -1,0 +1,7 @@
+package mx.yellowme.youst.dashboard.domain
+
+data class Dashboard(
+    val title: String,
+    val subtitle: String,
+    val options: List<DashboardOption>
+)
