+++ dashboard/src/main/java/mx/yellowme/youst/dashboard/domain/DashboardOption.kt
@@ -1,0 +1,15 @@
+package mx.yellowme.youst.dashboard.domain
+
+import java.util.*
+
+data class DashboardOption(
+    val id: String,
+    val title: String,
+    val subtitle: String,
+    val hexColor: String
+) {
+    val optionId: DashboardOptionId
+        get() {
+            return DashboardOptionId.valueOf(id.toUpperCase(Locale.ROOT))
+        }
+}
