+++ dashboard/src/main/java/mx/yellowme/youst/dashboard/domain/DashboardOptionId.kt
@@ -1,0 +1,8 @@
+package mx.yellowme.youst.dashboard.domain
+
+enum class DashboardOptionId {
+    CHALLENGES,
+    PLAYGROUND,
+    MINI_APPS,
+    PORTFOLIO
+}
