+++ dashboard/src/main/res/layout/dashboard.xml
@@ -6,30 +6,14 @@
     android:layout_height="match_parent"
     android:padding="@dimen/dimen_32">
 
-    <mx.yellowme.youst.core.components.YellowCircle
-        android:id="@+id/yellowCircle"
-        android:layout_width="@dimen/dimen_40"
-        android:layout_height="@dimen/dimen_40"
-        android:layout_marginBottom="@dimen/dimen_16"
-        app:layout_constraintBottom_toTopOf="@id/titleTextView"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <androidx.emoji.widget.EmojiTextView
-        android:id="@+id/titleTextView"
-        style="@style/Dashboard.Title"
-        android:layout_marginBottom="@dimen/dimen_8"
-        android:text="@string/dashboard_title"
-        app:layout_constraintBottom_toTopOf="@id/subtitleTextView"
-        app:layout_constraintStart_toStartOf="parent" />
-
     <TextView
         android:id="@+id/subtitleTextView"
         style="@style/Dashboard.Subtitle"
         android:layout_marginTop="@dimen/dimen_4"
         android:layout_marginBottom="@dimen/dimen_24"
-        android:text="@string/dashboard_subtitle"
         app:layout_constraintBottom_toTopOf="@id/guideline"
-        app:layout_constraintStart_toStartOf="parent" />
+        app:layout_constraintStart_toStartOf="parent"
+        tools:text="@string/dashboard_holder_subtitle" />
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guideline"
@@ -37,6 +21,14 @@
         android:layout_height="wrap_content"
         android:orientation="horizontal"
         app:layout_constraintGuide_percent="0.35" />
+
+    <androidx.emoji.widget.EmojiTextView
+        android:id="@+id/titleTextView"
+        style="@style/Dashboard.Title"
+        android:layout_marginBottom="@dimen/dimen_8"
+        app:layout_constraintBottom_toTopOf="@id/subtitleTextView"
+        app:layout_constraintStart_toStartOf="parent"
+        tools:text="@string/dashboard_holder_title" />
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/dashboardOptions"
@@ -47,8 +39,16 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/guideline"
         tools:itemCount="4"
-        tools:layoutManager="mx.yellowme.youst.core.hooks.recycler.SpanningGridLayoutManager"
+        tools:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
         tools:listitem="@layout/dashboard_option_item"
-        tools:spanCount="5" />
+        tools:spanCount="2" />
+
+    <mx.yellowme.youst.core.components.YellowCircle
+        android:id="@+id/yellowCircle"
+        android:layout_width="@dimen/dimen_40"
+        android:layout_height="@dimen/dimen_40"
+        android:layout_marginBottom="@dimen/dimen_16"
+        app:layout_constraintBottom_toTopOf="@id/titleTextView"
+        app:layout_constraintStart_toStartOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
