+++ dashboard/src/main/res/values/strings.xml
@@ -1,7 +1,9 @@
 <resources>
-    <string name="dashboard_title">Welcome to Youst!</string>
-    <string name="dashboard_subtitle">Now you control a yellow knight riding a flying ostrich or stork, from a third-person perspective.</string>
+    <string name="title_dashboard">Dashboard</string>
+
+    <string name="dashboard_holder_title">Welcome to Youst!</string>
+    <string name="dashboard_holder_subtitle">Now you control a yellow knight riding a flying ostrich, from a third-person perspective.</string>
+
     <string name="dashboard_option_item_holder_title">Dashboard Element</string>
     <string name="dashboard_option_item_holder_subtitle">This is a show case element description</string>
-    <string name="title_dashboard">Dashboard</string>
 </resources>
